Added /LTCG /INCREMENTAL:NO flags for MSVC to avoid warnings Added static_casts in strod.hpp and hash_map.hpp to remove MSVC warnings

+++ include/glaze/util/hash_map.hpp
@@ -186,7 +186,7 @@ namespace glz
             if constexpr (allow_hash_check) {
                ht.hashes[i] = hash;
             }
-            ht.table[hash % m] = i;
+            ht.table[hash % m] = static_cast<uint8_t>(i);
          }
 
          return ht;
